backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5

scikit-learn==1.5.2
numpy==2.1.3

python-docx==1.1.2
pdfplumber==0.11.4

httpx==0.28.1
openai==1.59.7
